Finish shadowseek-core (minus pipeline.rs); update deps., README
Finished all of the shadowseek-core crate (for now), except for the pipeline.rs file. (Will do that after the rest of the crates are fleshed out.) Updated all of the subcrates' Cargo.toml files with the proper workspace settings, and updated the dependencies for both shadowseek-core and shadowseek-extractor. Finally, updated the README with credits and comparison to the Duometer tool.